Fix for compiling with older XCode

+++ src/gui-osx/CreateFloppyController.m
@@ -54,8 +54,8 @@
 		cRet=CreateBlankImage_CreateFile(szPath, cTracks, cSectors, cSides);
 		if(cRet==TRUE)
 		{
-			NSInteger ret = NSRunAlertPanel(@"Hatari", @"Insert newly created disk in", @"Ignore", @"A:", @"B:");
-			if(ret!=NSAlertDefaultReturn)
+			int ret = NSRunAlertPanel(@"Hatari", @"Insert newly created disk in", @"Ignore", @"A:", @"B:");
+			if (ret != NSAlertDefaultReturn)
 			{
 				printf("%d\n",ret);
 				if(ret==-1) ret=1; //0=>Drive 0, -1=>Drive 1
